day9

+++ src/day2.rs
@@ -1,13 +1,13 @@
 use crate::vm;
 
 #[aoc(day2, part1)]
-fn day2_part1(input: &str) -> i32 {
+fn day2_part1(input: &str) -> i64 {
 	let mut state = vm::new_from_str(input).unwrap();
 	state.eval_0(&[12, 2]).unwrap()
 }
 
 #[aoc(day2, part2)]
-fn day2_part2(input: &str) -> i32 {
+fn day2_part2(input: &str) -> i64 {
 	let state = vm::new_from_str(input).unwrap();
 	
 	for range in 0 ..= 999_999 {
